requirements.txt
kafka-python
one_config
psycopg2
python-json-logger
pyyaml
requests
